properly adapt mappings to resolution changes
working. However, changing resolution sometimes triggers a crash in
TileOverlayView... to be investigated

+++ app/src/main/java/net/videosc/views/TileOverlayView.java
@@ -159,7 +159,7 @@ public class TileOverlayView extends View {
 					canvas.drawText(
 							String.valueOf(i + 1),
 							i % resolution.x * pixelSize.x + 3.5f * mApp.getScreenDensity(),
-							i / resolution.x * pixelSize.y + pixelSize.y - 3.5f * mApp.getScreenDensity(),
+							(float) (i / resolution.x) * pixelSize.y + pixelSize.y - 3.5f * mApp.getScreenDensity(),
 							mPaint
 					);
 				}
